Remove Python
This allows the BOSH release to be used without access to the public
Internet.  This does turn off some features of BIND; if those are needed
(they are difficult to enable / use in the release as-is), they can be
added back in with local copies of ply and other required Python
libraries.

Fixes #14